requirements.txt
python-docx
python-libnmap
